MainPage 컴포넌트에 있던 필터링 된 SNS 게시물 데이터를 요청해서 받아오는 로직을 훅으로 분리
- useFilteredSnsPosts hook
- 요청할 때 query에 담아 보내는 데이터는 static하게 작성한 상태 (테스트용)
